+++ src/Eve-O-Preview/Configuration/Implementation/ThumbnailConfiguration.cs
@@ -95,6 +95,7 @@ namespace EveOPreview.Configuration.Implementation
 			this.ClientLayout = new Dictionary<string, ClientLayout>();
 			this.ClientHotkey = new Dictionary<string, string>();
 			this.MinimizeAllClientsHotkeys = new List<string> { "Control+F22" };
+			this.ToggleCycleHotkeysHotkey = new List<string> { "Control+F23" };
 			this.DisableThumbnail = new Dictionary<string, bool>();
 			this.PriorityClients = new List<string>();
 
@@ -337,6 +338,8 @@ namespace EveOPreview.Configuration.Implementation
 		private Dictionary<string, string> ClientHotkey { get; set; }
 		[JsonProperty]
 		public List<string> MinimizeAllClientsHotkeys { get; set; }
+		[JsonProperty]
+		public List<string> ToggleCycleHotkeysHotkey { get; set; }
 		[JsonProperty]
 		private Dictionary<string, bool> DisableThumbnail { get; set; }
 		[JsonProperty]