+++ src/Eve-O-Preview/Presenters/Interface/IMainFormPresenter.cs
@@ -9,5 +9,7 @@ namespace EveOPreview.Presenters
 		void RemoveThumbnails(IList<string> thumbnailTitles);
 
 		void UpdateThumbnailSize(Size size);
+
+		void ApplyCycleHotkeysState();
 	}
 }